ScheduleMeeting/UI/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
